Relocate old langmuir source
